Define and use RequestTarget for fetch (#1231)
This does a little generic magic in order to allow for three basic forms
of request target natively:

* A simple string (this is the most costly, because we have to parse the
URL)
* A Url (in case one exists, as is the case with neqo-client)
* A three-tuple of things that turn into `str`

More of these can be implemented if that turns out to be convenient for
the caller.  Monomorphism ensures that only those forms that are used
are paid for.

This was a little fiddly to implement.  I would have liked to make the
string form more generic, ideally `AsRef<str>` but because we take
`&Url` and because `Url` implements `AsRef<str>` that causes rust to get
confused.  I tried fixing it, but you end up needing extra type
parameters on every invocation, which is less friendly than just having
people call `as_ref()` themselves.

Closes #1226.