Fast PTOs
Fastly have an option that enables faster PTOs.  This code adds a tuning
parameter for PTOs.  It allows for tuning the base scaling of the timer,
with a value of 10 being normal; 20 is half; 5 double.

The idea here is to allow people to set a value of ~20 to improve
latency in cases of tail loss.  The trade-off being that there are more
packets sent that might be redundant.  I think that we should use a
value of 20 in Firefox.

Draft.  Tests to come later.