Remove done() function.
The stream can change the state to the ending state during call to receive, header_unblocked and read_data. Also make sure that we handle the signal and remove the stream from the hash table.This makes the interface smaller and removes a bug, i.e. we did not remove the stream that was done after a header_unblocked call.