WebTransport negotiation
A generic negotiation structions, NegotiationState, is created. A feature can be in following states:
Disabled - it is disable for a connection.
Negotiating - it is enabled, but confirmation, i.e. a SETTING frame, from the peer is not received yet.
Negotiated and NegotiationFailed - these are the final states when SETTINGS frame from the peer is received.
When multiple features are available we may create another structure that holds all the features. Since only one feature is available at the moment, this is postponed.

NegotiationState is responsible to parse the SETTINGS and set the setting state for its feature. When a final state is set negotiation_done is called if a listener is supplied.

Currently the state of the WebTransport feature is considered when evaluating SETTINGS after a 0RTT. This is not by the spec and I will change it later.