WebTransport session
This PR creates a generic ExtendedConnectFeature structure that is responsible for a feature negotiation and keeps a HashMap of the extended connect requests. The HashMap will be used for finding active sessions.

An extended connect request will be created as a regular fetch request only the listener for the events will be different.  A generic ExtendedConnectSession structure is created for each extended connect request and is responsible for listening to request events and following the state of the request/session, e.g. the session is being negotiated, a stream is closed, etc..This structure will be responsible for bookkeeping, i.e. it will hold information about active streams open on a session and it will make sure to close all streams if the session request is closed.
Currently ExtendedConnectSession cannot read data from the request stream, this will be fixed in a separate PR.
On the server side, upon receiving headers a regular http request will be transformed into an extended connect request and a new ExtendedConnectSession will be set to be the listener for the request.

Regular http request functions will be used for closing an extended connect request, e.g. steam_stop_sending, cancel_fetch, etc.

Additional changes:
Now it is not easily possible to distinguish between Http stream and Push stream, therefore  RecvMessage holds the type of the stream.
WebTransport’s extended connect request will be created as a normal fetch request. The listener for the events on this request will be different

All WebTransport events are put into a separate enum structure.