WebTransport streams
This implements WebTransport streams. A stream is a very thin layer that is responsible for posting RecvStreamEvents and SendStreamEvents and removing itself from the ExtendedConnectSession.

The WebTransport streams reuse a lot of API that was used for HTTP requests, e.g. reading data, stream_close_send, stream_reset. Also all events are reused. The only difference between a HTTP request and WebTransport is that WebTransport starts with a new event “NewStream”(if they are peer initiated) and the HTTP request starts with HeadersReady. The events now use Http3StreamInfo that contains the type and the stream number.

Parts that are not so nice is that Http3Connection cannot create streams directly because it need Box<dyn Send/RecvStreamEvents>, so the WebTransport streams are partially handle by Http3Connection, i.e. the are rejected if a session does not exist, and then they are actually created by Http3Client. (This is a TODO to improve)

On the server side stream handles are renamed from ClientRequestStream to Http3OrWebTransportStream, because they are used for HTTP3 request and WebTransport streams.

Additional changes:
- Rename read_resoonse_data/read_request_data/send_request_data into read_data/send_data. This is now a more general function used for HTTP requests and WebTransport streams.
- Http3Connection now knows its Role.
- Set self.needs_processing to true in the Http3ServerHandler. This flag is  used to make sure that a particular server is processed(function process() is called) after some API activities. Some of this API may trigger the process in a different way, but it is better to be sure. This is server side only, we may consider improving the server side at some point in the future (it is not a priority at the moment).
- Clippy also ask for changing Http3Server::process_http3() because it is too long.