Encode headers when they are set.
The function encode_header_block does not produce any errors that need to be picked up by the main processing loop. This was change in the previous change set. Therefore the encoding can be performed immediately.

This PR will change current behavior regarding the order in which the encoder instructions and header blocks are sent.