Implement sending of 1xx responses and refactor SendMessage
The biggest change is that SendMessage uses BufferedStream and does not need states. The header blocks are added using send_headers and they are buffered into the BufferedStream. There may be multiple header blocks, but the code does not enforce or check that only server can send multiple headers blocks or that the multiple 1xx header blocks are followed by a non-1xx header block. This is left for the application to enforce.

The server and client API are unified, both sides use send_header, send_data  and close_send functions.