Ensure that http messages that are sent have a propriete form.
This change adds a state machine for http meassages.
A HTTP message, request and response, consists of headers, optional data and an optional
trailer header block and has appropriate states: WaitingForHeaders, WaitingForData, TrailersSet and Done.
This state machine does not reflect what was already sent to the
transport layer but only reflect what has been supplied to the `SendMessage`.