Make server work with ZeroRtt
The server side did not know about ZeroRTT state. It was working because the server API was not strict about the state of the connection when writing data into a stream. As the last changes unified the API, the server started to trigger assertion, i.e. the  http3 connection was in the initialized state and in that state it is not possible to write data.

The patch sets ZeroRtt state on the server side as well.