Improve closing of WebTransport as well as creating it
Create enum SessionCloseReason with values: Error, Status and Clean.
StreamIId of the ExtendedConnectSession stream can be known before creating a stream. To make this work we need to split the fetch function into: 1) creating a transport stream and 2) initiating the HTTP stream. This simplifies ExtendedConnectSession states. This change was caused by removing negotiation_done. On the server side negotiation_done, it was only use for adding the stream_id to thee ExtendedConnect Session. On the client side this is renamed to headers_ready and all logic from headder_ready function and negotiation_done was moved to a single function which makes the code cleaner.