FrameReader needs a more generic connection reader.
HTTP/3 frame read data directly from a transport connection.WebTransport frames are one level above. These frames are sent as the data body of a CONNECT stream. Therefore FrameReader must be able to read from a transport connection and from a RecvStream.

The test for RecvStream will be added when the frame reading functionality is added to the WebTransport feature.