Make WebTransportSession as a send and recv stream that holds Recv/SendMessage.
ExtentedConnectSession was renamed to WebTransportSession. This was not a proper generalization of extended connections. Additional extended connect will need a separate implementation, e.g. WebTransport. We should make a generalization as needed.  Something like Masque has a potential to reuse functionality of WebTransportSession, but let’s wait for an implementation to make WebTransportSession generic.

In the current implementation ExtenedConnectSession was only a listener to the underlying http stream (Recv/SendMessage). This was working fine until the need for reading the body of the http stream was introduced. Since listeners cannot read data from a http stream, the task was done by the Htttp3Connection(e.g. function read_webtransport_control_stream).
This patch creates WebTransportSession’s Recv/SendStream that holds the http Recv/SendMessage. All reads and writes are done in the WebTransportSession and Http3Connection does not need to know about WebTransport details. There are no more special reads for the WebTransport streams.