zero-length unknown frame payloads (#1315)
* Try a different approach for priority_update test failures

* zero-length unknown frame payloads

Don't try to read this, it doesn't work.

The real fix is in neqo-http3/src/frames/reader.rs

This includes a few other changes:

* I made the incremental ignoring decoder not work for zero-length
things.  This is the source of the problem.  Fixing this doesn't help in
this case, but it should block similar mistakes.

* Enable RUST_LOG for running in CI.  The test runner now captures
output, so we are able to get per-test logs when they fail.

* Use || instead of | for some boolean condition.

* Fixed a typo in comments/logs.

* Improved exchange_packets.

* Streamlined the event handling in priority_update.