Use map for encoding resumption token
I was going to just commit this and push directly as it is trivial, but
I wanted to share...

Clippy generated an error with this code, suggesting that it use
`.map_or(None, |x| ...)`.  Then it suggested that that
was better replaced with `.and_then(|x| ...)`.  Finally, it suggested
that a simple `.map(|x| foo)` was better than the `.and_then(|x|
Some(foo))` there.

In any case, here we are after three separate clippy lints.