Fix changes to VN
CI not working really hid a major problem here.  Let's not let this rot
so much again...

The code didn't work here because we weren't saving the wire version
when we parsed the packet header for a packet that had an unsupported
version.  Now we do, so we can correctly generate a VN packet.

I took the liberty of renaming as_u32 to wire_version so that we are
more consistent with the naming of the version functions.