Don't send small Initial packets (#1337)
* Don't send small Initial packets

This turned out to be a little fiddly, but it should also improve
handshake reliability (in rare scenarios, admittedly).

Changes:

1. At the client, discard the Initial packet number space when the
Handshake keys are installed.

2. Because Initial packets from the client are not sent in its second
flight, the server won't get an ACK, which otherwise drives the RTT
estimate.  Rather than disrupt all the tests, when the server discards
its Initial packet number space it will add an implicit RTT sample.

Note: The server will not increase its congestion window as a result of
this.  We might consider doing something about that later.

3. Because the client might only receive Initial packets from the
server, if it discards the Initial packet number space there will be
nothing to send.  No Initial packets and no Handshake packets, not even
acknowledgments.  So, to prevent deadlocks, if *either* endpoint
receives a packet from a space in which it has discarded keys, it will
probe in the next available space when it next sends something.  It will
only do this prior to connection establishment, meaning that it won't
probe with Application packets.

Closes #1336.

* New test for probes after confirmed; tweaks