Fix detection whether packet was sent after recovery start
Before this patch, the congestion window was reduced multiple times per
round trip time (RTT) on multiple packet loss. Not it is capped to one
per RTT.

The recovery_start variable is only updated after transitioning from
RECOVERY_START to RECOVERY when sending out packets after the recovery
event. So if the state is RECOVERY_START (transient), all packets are
from before the RECOVERY. Therefore we must only check the
recovery_start time after trasitioning to the RECOVERY state.

If we haven't encountered a packet loss, the packets always count as
after recovery_start.

This is to fix the congestion algorithm to be [spec aligned][1]:

> The recovery period aims to limit congestion window reduction to once
> per round trip. Therefore, during a recovery period, the congestion
> window does not change in response to new losses or increases in the
> ECN-CE count.
>
> A recovery period ends and the sender enters congestion avoidance when
> a packet sent during the recovery period is acknowledged. This is
> slightly different from TCP's definition of recovery, which ends when
> the lost segment that started recovery is acknowledged [RFC5681].

This was found when investigating [Bug 1852924][2], but doesn't fix the
upload speed issue.

[1]: https://www.rfc-editor.org/rfc/rfc9002.html#section-7.3.2
[2]: https://bugzilla.mozilla.org/show_bug.cgi?id=1852924