Cap the PTO count when working out the period
The previous code would overflow and produce a 0ns timeout when the PTO
count got large (well over 24).  This is because the `fast_pto` value
tends to have trailing zero bits and shift values between 32-(number of
zeros) and 31 (inclusive) would still shift left.

This code instead stops increasing the PTO once it hits 24.  That should
be enough for most cases.

Closes #1602.