Switch to rust nightly for CI
And fix the pre-commit hook to check using nightly.

This is because we now rely on rust nightly for some rustfmt
configuration options.  This will make our canary builds a little more
forward-looking which comes with a little less stability, but I think
that's better than adding a specific build just for formatting.