fix(client): wait for resumption token (#1676)
The QUIC Interop testcase `resumption` requires a client to download two files
on two consecutive connections, establishing the second connection with the
resumption token of the first.

Previously the client would close the first connection once the first file is
downloaded. There is a race condition where the first file might be finished
downloading before receiving the resumption token from the server.

With this commit, the client will wait for both (1) the file being downloaded
and (2) receiving the resumption token from the server.