fix(tests/cc): compare cwnd before and after congestion (#1694)
The `cc_slow_start_to_cong_avoidance_recovery_period` test sends two flights of
data from client to server. In the second flight the first packet is dropped,
which is thus not included in the second ACK of the server to the client.

Due to the dropped packet, the client is expected to move into recovery state.
But the test only checks the largest acknowledged packet. It does not assert the
state change. Making `ClassicCongestionControl::on_congestion_event` a no-op
does not fail the test.

With this commit, the test asserts that the congestion window decreases, thus
presuming a state change.