refactor(client): de-duplicate process & run and split into h3 & h09  (#1727)
* refactor(bin): move client.rs to client/main.rs

As a preparation to introducing sub-modules to `client`.

Done in a separate commit for git to recognize the move as a move, not as a
remove and add.

* refactor(client): de-duplicate process & run and split into h3 & h09

The Neqo Client binary supports both http3 and http09 (prev. "old").

Before this commit both the http3 and the http09 implementation had their own
`run` and `process` `fn`, orchestrating the interaction between handler, client
and I/O. While similar, they had subtle differences e.g. when to terminate.

This commit splits the http3 and http09 specific logic into two separate modules,
but extracts duplicate logic (e.g. `run` and `process`) into the shared root module.

---------

Co-authored-by: Lars Eggert <lars@eggert.org>