ci(interop/action): always upload comment data (#1790)
In https://github.com/mozilla/neqo/pull/1785 the QUIC Network Simulator workflow
failed.

https://github.com/mozilla/neqo/actions/runs/8523273988/job/23345192160?pr=1785

This triggers the QUIC Network Simulator Comment workflow.

https://github.com/mozilla/neqo/actions/runs/8524906268

Though currently, the former does not upload the _comment_ artifact on failure,
which is to be consumed by the latter.

This commit makes the former always upload the comment data, such that the
latter can consume it on failure.