fix(bin/bench): benchmark handshakes per second (HPS) without zero RTT (#1795)
The "100-seq-conn/1-1b-resp (aka. HPS)" benchmark supposedly measures the time
for 100 uniform handshakes. Though `neqo-client` by default uses a resumption
token for zero RTT if available. Thus the first is a normal and all following 99
handshakes are a zero RTT handshake.

To simplify the benchmark and to measure the same handshake type, establish a
single connection within the benchmark function and have criterion do all the
iterating.