fix(bin): don't sleep when events available
A call to `process_*` can emit events as a side effect, e.g. a
`ConnectionEvent::StateChange(State::Connected)`. These are not returned from
the function call, but instead internally enqueued to be later on consumed
through the various `Provider::next_event` implementations.

https://github.com/mozilla/neqo/blob/166b84c5a3307d678f38d9994af9b56b68c6b695/neqo-common/src/event.rs#L9-L15

In the case of `neqo-client` the events are consumed through
`self.handler.handle` which internally calls `Provider::next_event`.

A client or server should not go to sleep, waiting for either a UDP datagram to
arrive or a timeout to fire, when there are events available.

This commit ensures `ready().await` is only called when no events are available.