fix(server): log error when failing to read file (#1788)
* fix(server): log error when failing to read file

As part of the QUIC Interop testcases, a server needs to be able to respond to a
request with the content of a local file.

Previously, when failing to read a file, the server would simply reply with the
predefined `SELF::MESSAGE`.

Now the server logs the error and moves on.

* Send 404