fix(bin): consider env when no verbosity flag is set (#1848)
* fix(bin): consider env when no verbosity flag is set

There are two ways to specify the maximum log level of `neqo-server` and
`neqo-client`. Via the `RUST_LOG` environment variable and since
https://github.com/mozilla/neqo/pull/1692 via the verbosity flags (e.g. `-v`).

Previously, if no verbosity flag was provided, `INFO` was set as the maximum log
level, the `RUST_LOG` environment variable was simply ignored.

With this commit, if no flag is set, the `RUST_LOG` environment variable is
considered, and only then the `env_logger` crate default value (`ERROR`) is used.

In other words, the precedence order now is:

1. command line flags (e.g. `-v`)
2. `RUST_LOG` environment variable
3. default `ERROR` level

* clippy

* Move `verbose` into `SharedArgs`