fix(bin/client): don't close closing connection (#1866)
* fix(bin/client): don't close closing connection

The `bin/src/client/mod.rs` `Runner::run` function continuously checks whether
there is more work. In case there is none, it initiates closing of the
connection (`self.client.close`) and then `continue`s to the top of the loop in
order to send out a closing frame.

https://github.com/mozilla/neqo/blob/14cafbaa7fa88434def2c1d19e932c08e00173f8/neqo-bin/src/client/mod.rs#L376-L409

There is a potential busy loop when closing an already closing connection.
`Runner::run` will call `self.client.close` and then continously `continue` to
the top of the loop.

This commit differentiates a connection state in `NotClosing`, `Closing` and
`Closed`. It only attempts to close a `NotClosing` connection and only then
`continue`s to the top of the loop.

* Introduce ConnectionError::is_error and implement TryFrom