ci: rename .taskcluster to taskcluster
While we'd like to keep the CI configuration files hidden in this repo,
I came across a bug in Taskgraph triggered by changing the root
Taskgraph directory to a non standard location. This bug will require a
new release to fix (and I suspect there will be other bugs revealed
afterwards).

In the meantime, we can move the files to the standard place while this
is fixed. I filed https://github.com/taskcluster/taskgraph/issues/522 to
track the fix.