revert: remove ttl information from datagram
https://github.com/mozilla/neqo/pull/1568 introduced TTL information to
datagrams. On the input path it would take the ttl information, but not act on
it. On the output path it would set only "the default TTL on many OSes".

https://github.com/mozilla/neqo/blob/66504908e5fa070a8a5fa67d8b5a201d2c9a5cc5/neqo-transport/src/path.rs#L576

This commit removes the ttl information from `Datagram`, thus reverting a subset
of https://github.com/mozilla/neqo/pull/1568.

This is partially motivated by https://github.com/mozilla/neqo/pull/1920,
introducing `quinn-udp` as the IO library of choice, which does not support
reading and writing TTL.

See also discussion in
https://github.com/mozilla/neqo/pull/1920#discussion_r1644113749.