chore(neqo-udp): drop tokio dependency
`neqo-udp` is used in Firefox. Firefox uses `cargo vet` to audit its
dependencies. `tokio`, the dependency of `neqo-udp`, is not audited as
`safe-to-deploy`. `cargo vet` will require `safe-to-deploy` for `tokio` even
when behind a feature flag.

To work around this limitation in `cargo vet`, drop `tokio` dependency in
`neqo-udp`, moving `tokio` specific logic into `neqo-bin`.

See details in:

- https://github.com/mozilla/cargo-vet/issues/626
- https://bugzilla.mozilla.org/show_bug.cgi?id=1907810
- https://phabricator.services.mozilla.com/D212959#inline-1194604