fix(benches/transfer): remove throughput
The `neqo-transport/benches/transfer.rs` benchmarks use the
`test-fixture/src/sim` simulator. The simulator can travel in time, i.e. it
simulates time.

The _wall-clock time_ of a single benchmark run is not the amount of time it took to
transfer `TRANSFER_AMOUNT`. The _simulated time_ is the amount of time it took
to transfer `TRANSFER_AMOUNT`.

`criterion` will use the _wall-clock time_, not the _simulated time_ to
calculate the throughput based on `TRANSFER_AMOUNT`. The resulting throughput
number is not meaningful.

This commit removes the call to `group.throughput`, thus removing the misleading
`criterion` throughput reporting.