docs(transport): fix old/new mixup
In the `path_forwarding_attack` unit test

``` rust
/// This simulates an attack where a valid packet is forwarded on
/// a different path.  This shows how both paths are probed and the
/// server eventually returns to the original path.
fn path_forwarding_attack() {
```

data is send on two paths, a IPv6 path (aka. "old") and an IPv4 path (aka. "new").

Towards the end of the test it asserts one path, but mentions the other in the
doc comment.

This commit fixes the mix-up in the doc comment.