fix(udp): ignore empty datagram (#2044)
* test(udp): assert ignoring of empty datagram

* fix(udp): ignore empty datagram

When receiving an emtpy datagram `meta.len` and `meta.stride` would be `0`.
Chunking the receive buffer via `.chunks(0)` would panic with:

```
chunk size must be non-zero
```

See also panic docs on `slice::chunks`:

https://doc.rust-lang.org/std/primitive.slice.html#method.chunks

With this commit `recv_inner` ignores the empty datagram.

In addition, under the assumption that an empty datagram is not a faulty event,
`recv_inner` attempts another receive call on the socket. This ensures that the
socket eventually returns `WouldBlock` and is thus registered for the next
wake-up with the corresponding event loop (e.g. tokio`).