fix: Single-packet RTX in Initial/Handshake + RTX handshake ACKs (#2027)
* fix: Send ACKs during handshake after they are lost

This is not yet complete. It works for the first RTX of a lost server
Initial, which now contains the same data as the original
transmission. A third RTX only contains pings. Something is off about
our probe logic. Will continue investigating, but want to see what
the bench and QNS impacts of this change are.

* Improvements

* Remove debug output

* Suggestion from @mxinden

* doc comment

* Check if `L1`/`C1` improve if this is also done in the Handshake space

* Update comment

* ACK more aggressively during the handshake

* Undo

* Update neqo-transport/src/tracking.rs

Signed-off-by: Lars Eggert <lars@eggert.org>

* Update neqo-transport/src/recovery/mod.rs

Signed-off-by: Lars Eggert <lars@eggert.org>

---------

Signed-off-by: Lars Eggert <lars@eggert.org>