perf(transport): don't pre-allocate mtu on max_datagram_size
The `neqo_transport::Connection::max_datagram_size` creates an `Encoder`, writes
a packet header and a packet number and determines how many bytes of the mtu are left.

https://github.com/mozilla/neqo/blob/28f60bd0ba3209ecba4102eec123859a3a8afd45/neqo-transport/src/connection/mod.rs#L3408-L3427

The `Encoder` only has to hold the packet header and the packet number. Yet it
is initialized with `Encoder::with_capacity(mtu)`.

https://github.com/mozilla/neqo/blob/28f60bd0ba3209ecba4102eec123859a3a8afd45/neqo-transport/src/connection/mod.rs#L3408

Note that `PacketBuilder::short` and `PacketBuilder::long` read the
`Encoder::capacity` through `PacketBuilder::infer_limit`. But
`PacketBuilder::infer_limit` falls back to `2048` if the capacity is below `64`,
which will be the case when using `Encoder::default()` instead of
`Encoder::with_capacity(mtu)`. `2048` should be plenty enough for the packet
header and the packet number.

https://github.com/mozilla/neqo/blob/28f60bd0ba3209ecba4102eec123859a3a8afd45/neqo-transport/src/packet/mod.rs#L152-L180

https://github.com/mozilla/neqo/blob/28f60bd0ba3209ecba4102eec123859a3a8afd45/neqo-transport/src/packet/mod.rs#L188-L225

https://github.com/mozilla/neqo/blob/28f60bd0ba3209ecba4102eec123859a3a8afd45/neqo-transport/src/packet/mod.rs#L135-L141

This commit prevents the wasted allocation by using `Encoder::default()` instead
of `Encoder::with_capacity(mtu)`.