fix: More migration tests and related fixes (#2170)
* test: More migration testing

This expands the migration test quite a bit. It now tests alll the way
until retirement of the original path, for both port-only
(`rebinding_port`) and address-and-port (`rebinding_address_and_port`)
changes.

`rebinding_address_and_port` is succeeding, but `rebinding_port` is
currently failing. That's because we treat it differently for some
reason. If we replace `Paths::find_path_with_rebinding` with
`Paths::find_path`, i.e., do proper path validation when only the port
changes, the test succeeds.

Leaving this out in case I'm missing something about the intent of the
difference.

* fix: Replace `find_path_with_rebinding` with `find_path`

We need to do a path challenge even if only the remote port changes.

* Remove unused arg from `received_on`

* Fixes

* identity

* More review comments

* Simplify

* Fix merge

* Address review comments

* Migrate to preferred address

* Fix URL splitting

* Restore `find_path_with_rebinding`

* Fix test for not doine path challenge if only port rebinds

* Revert "Fix test for not doine path challenge if only port rebinds"

This reverts commit 97fafa6d2e7f04e1011af0eb308f55f4dbf97a4c.

* Revert "Restore `find_path_with_rebinding`"

This reverts commit 302961f3615a991953e9cbeda6f04a266ae8c910.

* Remove now-unused test names

* fmt

---------

Signed-off-by: Lars Eggert <lars@eggert.org>