deps: pin url to v2.5.1
mozilla-central depends on `url` `v2.5.1`:

https://searchfox.org/mozilla-central/rev/6050bf4eca89956c9d91bfd89fa59294ae32a689/Cargo.lock#6715-6725

The latest version of `url` is `v2.5.3`, which cargo automatically updates to,
given that `Cargo.lock` is not checked in.

`url` v0.2.5.3` introduced the `std` feature which breaks CI.
https://github.com/mozilla/neqo/pull/2219 enabled the feature, unbreaking CI,
but breaking mozilla-central, given that mozilla-central still uses `url`
`v2.5.1` which does not have the the `std` feature.

This commit pins `url` to `v2.5.1` instead and removes the `std` feature.